Modified Get{BYTE,WORD,DWORD}() to return the respective type
Having e.g. GetWORD() return a WORD rather than int eliminates the need for
a cast when assigning to WORD variables. We use casts only when assigning
to a different (but same-sized) type.

Also, check for EOF explicitly when calling fgetc().

+++ solvespace.h
@@ -373,9 +373,9 @@ public:
     Vector      origin, u, v;
 
     int Getc(void);
-    int GetBYTE(void);
-    int GetWORD(void);
-    int GetDWORD(void);
+    BYTE GetBYTE(void);
+    WORD GetWORD(void);
+    DWORD GetDWORD(void);
 
     void LoadGlyph(int index);
     bool LoadFontFromFile(bool nameOnly);
